Update method stored in "call" for recombineSL and recombineCVSL
The new method set for recombineCVSL was not being passed to the "call" object, which caused the risks to be calculated incorrectly when the new CV SL object was passed to summary.CV.SuperLearner. The edits to recombineCVSL fixes this.

Risks were calculated correctly for new SL objects created from recombineSL, and the edits to recombineSL only update the call object to reflect the new method used.